Source code synced with Eugene game development kit 8.7.8

+++ source/blackgdk.cpp
@@ -2192,7 +2192,7 @@ namespace BLACKGDK
    return this->is_software_render();
   }
 
-  unsigned long int Screen::get_color() const
+  unsigned int Screen::get_color() const
   {
    return this->get_depth();
   }
